+++ src/main/java/net/cryptic_game/server/server/http/HttpEndpoint.java
@@ -1,4 +1,4 @@
-package net.cryptic_game.server.http;
+package net.cryptic_game.server.server.http;
 
 import org.json.simple.JSONObject;
