+++ src/main/java/net/cryptic_game/server/server/http/endpoints/PlayerLeaderboardEndpoint.java
@@ -1,12 +1,7 @@
-package net.cryptic_game.server.http.endpoints;
+package net.cryptic_game.server.server.http.endpoints;
 
-import net.cryptic_game.server.http.HttpEndpoint;
-import net.cryptic_game.server.microservice.MicroService;
-import net.cryptic_game.server.utils.JSONBuilder;
+import net.cryptic_game.server.server.http.HttpEndpoint;
 import org.json.simple.JSONObject;
-
-import java.util.Collections;
-import java.util.UUID;
 
 import static net.cryptic_game.server.socket.SocketServerUtils.sendHTTP;
 import static net.cryptic_game.server.socket.SocketServerUtils.sendRaw;
