+++ src/main/java/net/cryptic_game/server/server/http/endpoints/PlayersOnlineEndpoint.java
@@ -1,7 +1,7 @@
-package net.cryptic_game.server.http.endpoints;
+package net.cryptic_game.server.server.http.endpoints;
 
 import net.cryptic_game.server.client.Client;
-import net.cryptic_game.server.http.HttpEndpoint;
+import net.cryptic_game.server.server.http.HttpEndpoint;
 import org.json.simple.JSONObject;
 
 import static net.cryptic_game.server.utils.JSONBuilder.simple;
